refactor: fold MCP App support into ground_location_tool
Drops ground_location_app_tool. Existing ground_location_tool now emits
meta.ui + inline rawHtml using a shared renderGroundLocationAppHtml,
extracting origin + nearby_pois from its rich composite output.
